Extract mutation list from a directory of constellation files.

"Constellations" are files produced from https://github.com/cov-lineages/constellations, which represent the cov-lineage team's best knowledge about which mutations define a lineage. See details.

Usage

astronomize(path = NULL)

Arguments

path

Path to the constellations folder in the cov-lineages/constellations repository. If NULL, built-in definitions are used. There is no need to clone the repo unless you want updated definitions.

Value

A lineage definition matrix for use with provoc.

Details

The constellations repo is no longer updated to new variants.

Code is adapted from scripts/estimate-freqs.R in https://github.com/PoonLab/gromstole.

From the repo, a constellation "a collection of mutations which are functionally meaningful, but which may arise independently a number of times".

Updated as of version 0.5.0: constellation files are hardcoded into provoc, there is no longer a need to clone the repository.

If a path is provided, the function uses that repository. This is useful if, say, you clone a historical version of the repo to ensure that you'r eworking with definitions that were known at the time.

See also

Examples

# After cloning the constellations repo
lineage_defs <- astronomize(path = "../constellations")
#> Warning: Path does not exist. Using built-in definitions.
dim(lineage_defs)
#> [1]  37 325
lineage_defs <- filter_lineages(lineage_defs, c("B.1.1.7", "B.1.617.2"))
dim(lineage_defs) # rows and columns have changed
#> [1]  2 36