Ensure mutations are present in the data and lineage_defs

Finds the intersection between the mutations present in coco and lineage_defs. Will squash lineages together if the resulting mutation list is too similar (see details).

Usage

fuse(coco, lineage_defs, min_perc = 0.01, verbose = FALSE)

Arguments

coco

A data frame with a column labelled mutation.

lineage_defs

Rownames are lineages, column names are Mutations.

min_perc

A lineage must have at least min_perc of the mutations.

vebose

Print information about mutations that were removed by the fusion. 0 (FALSE) returns errors, 1 (TRUE) returns warnings and some info about relative mutation counts, and 2 returns all mutations in each.

Value

A data frame with the same columns as coco (possibly fewer rows) and the same columns plus new columns for the lineages. The provoc function expects this structure.

Details

First, the intersection of the mutations is found.

The columns of lineage_defs are subsetted according to this intersection. If this removes all mutations for any lineage, that lineage is removed from the study (a warning is given if the lineage was a pre-specified voc).

After removing mutations, it's possible that some rows lose their distinctive mutations and become identical. In this case the names of the lineages are pasted together and only one of the rows are kept.

Duplicate mutation names in coco are NOT removed. It is safe to use this function on a data frame that contains multiple samples.

Examples

data(Baaijens)
b1 <- Baaijens[Baaijens$sra == Baaijens$sra[1], ]
b1$mutation <- parse_mutations(b1$label)
lineage_defs <- astronomize() |>
    filter_variants(c("B.1.1.7", "B.1.429", "B.1.617.2"))
#> Error in filter_variants(astronomize(), c("B.1.1.7", "B.1.429", "B.1.617.2")): could not find function "filter_variants"
fused <- fuse(b1, lineage_defs)
#> Error in rownames(lineage_defs): object 'lineage_defs' not found
head(fused)
#> Error in head(fused): object 'fused' not found